Add collect plugins
